Async Icons
Asynchronously set an icon and handle errors.

setIconAsync takes an icon which you want to change, a Gio.File representing an image file(local or remote) that may or may not exist, and a fallback icon name to use if the file does not exist. It then asynchronously loads the contents into a gicon if successful otherwise it sets the icon to the fallback icon name.

This removes the need for blocking I/O when loading images from a file.

(cherry picked from commit f6535d451cf02a5b4cc520ed842e64f95649868e)